Move calibration window to its own module
